:recycle: 커스텀 랜덤 디펜스의 테스트 코드에서 세부 기록까지 CustomDefenseRecordTest에서 검증하도록 수정
